misc changes

+++ isperfect_v2.py
@@ -4,7 +4,7 @@ import amath
 
 if len(argv) > 1:
     if argv[1].isdigit():
-        num = int(argv[1])
+        num = long(argv[1])
         n = "is not"
         perf = amath.is_perfect_v2(num)
         if perf:
